Schema logging strongly improved + log config
The dynamic schema loading tended to be touchy. When a module had a compilation error, it would often be silently ignored, resulting, at best, in incomplete GraphQL schemas.
I have added lots of logs in DEBUG and a sample entry in the settings file to display them.
Also took the opportunity to improve the sample logging configuration to make debugging easier and show what can be done.